// components/shared/navigation/FileTree/types.d.ts
import type { MutableRefObject } from "react";

export interface TreeItem {
  id: string;
  label?: string;
  fileUrl?: string;
  children: TreeItem[];
  type?: "folder" | "file";
  collapsed?: boolean;
}

export type TreeItems = TreeItem[];

export interface FlattenedItem extends TreeItem {
  parentId: null | string;
  depth: number;
  index: number;
}

export type SensorContext = MutableRefObject<{
  items: FlattenedItem[];
  offset: number;
}>;